backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.0.1
aiosqlite==0.20.0
apscheduler==3.10.4
email-validator==2.2.0
